Make DurableCursor and MongoProxy more alike; try disconnecting
Make the messages logged by DurableCursor and MongoProxy more
consistent.

Make DurableCursor sleep between attempts for a maximum of five
seconds, like MongoProxy does.

For both DurableCursor and MongoProxy, if AutoReconnect fails after 60
seconds, then call disconnect() on the connection and then try for
another 60 seconds. Here's the rationale for this... We've seen
circumstances in which PyMongo simply refuses to successfully
reconnect, even after many minutes, when the replicaset has come back
up and is perfectly healthy. We're not sure what's causing this, but
we're hoping that the PyMongo documentation doesn't lie when it claims
that calling disconnect() will completely close all connections to the
server(s), and we're _also_ hoping that whatever problem is causing
the reconnect attempts to fail will be cleared up by calling
disconnect() and forcing PyMongo to reconnect from scratch.